knowledge_service/requirements.txt
# ── Web 框架 ──────────────────────────────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# ── ORM & 数据库 ──────────────────────────────────────────
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0
alembic>=1.13.0

# ── 数据校验 & 配置 ──────────────────────────────────────
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# ── 对象存储 ──────────────────────────────────────────────
minio>=7.2.0

# ── HTTP 客户端 ───────────────────────────────────────────
httpx>=0.25.0

# ── 文档解析 ──────────────────────────────────────────────
PyPDF2>=3.0.0
python-docx>=1.0.0

# ── LLM (阿里云 DashScope) ───────────────────────────────
dashscope>=1.14.0

# ── JSON 容错修复（兼容 LLM 输出格式偏差）──────────────
json-repair>=0.28.0

# ── 异步文件 IO ──────────────────────────────────────────
aiofiles>=23.2.0
